requirements.txt
fastapi==0.110.0
uvicorn==0.15.0
SQLAlchemy==2.0.29
asyncpg==0.29.0
pydantic==1.10.14
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
prometheus-fastapi-instrumentator==7.0.0
